add popover component
