+++ pages/transactions/Transactions.module.css
@@ -258,186 +258,6 @@
   margin: 0px;
 }
 
-.popover-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: left;
-  margin-top: 1%;
-  border: 2px solid rgba(234, 234, 234, 0.4);
-  border-radius: 8px;
-  width: 370px;
-  padding-left: 30px;
-  padding-right: 30px;
-  padding-bottom: 30px;
-  overflow: hidden;
-}
-
-.popover-div {
-  display: flex;
-  flex-direction: column;
-}
-
-.popover-header {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.add-title {
-  color: #253c85;
-  margin-bottom: 0px;
-}
-
-.x-button {
-  width: 25px;
-  height: 25px;
-  cursor: pointer;
-  margin-top: 15px;
-}
-
-.x-button:hover {
-  opacity: 0.5;
-  transition-duration: 0.2s;
-}
-
-.success-div {
-  margin-top: 20px;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: center;
-  align-items: center;
-  padding: 15%;
-}
-
-.success-title {
-  color: var(--noirish);
-  font-style: normal;
-  font-weight: bold;
-  font-size: 24px;
-  letter-spacing: 2px;
-  line-height: 31px;
-  text-transform: none;
-  margin: 0px;
-}
-
-.success-message {
-  text-align: center;
-  margin-bottom: 15%;
-}
-
-.success-close-button {
-  color: #525454;
-  border: 1.3px solid #525454;
-  border-radius: 7px;
-  font-family: Avenir;
-  font-style: normal;
-  font-weight: 800;
-  font-size: 14px;
-  line-height: 19px;
-  text-transform: none;
-}
-
-.add-more-button {
-  background-color: var(--sapphire);
-  color: white;
-  border-radius: 7px;
-  font-size: 14px;
-  margin-left: 20px;
-  text-transform: none;
-}
-
-.select-category {
-  font-family: Avenir;
-  font-style: normal;
-  font-weight: 800;
-  font-size: 12px;
-  line-height: 16px;
-  text-transform: uppercase;
-}
-
-.accordion {
-  background-color: #f7f7f7;
-}
-
-.accordion-open {
-  background-color: #eaeaea;
-}
-
-.drop-triangle {
-  width: 10px;
-  height: 5px;
-}
-
-.autocomplete-option-select:hover {
-  background-color: var(--sapphire);
-}
-
-.autocomplete-text-field {
-  background-color: white;
-}
-
-.autocomplete-input {
-  font-size: 12px;
-  line-height: 20px;
-}
-
-.amount-action {
-  display: flex;
-  width: 100%;
-  flex-direction: row;
-  margin-top: 10px;
-  gap: 20px;
-}
-
-.amount-field {
-  width: 90%;
-}
-
-#amount {
-  flex: 2;
-}
-
-#action {
-  flex: 4;
-}
-
-.radio-label {
-  font-style: normal;
-  font-weight: normal;
-  font-size: 12px;
-  line-height: 16px;
-
-  color: #646464;
-}
-
-.message-field {
-  width: 100%;
-  background-color: #f7f7f7;
-}
-
-.message-field-input {
-  font-size: 12px;
-  line-height: 20px;
-  margin-left: 15px;
-  margin-right: 15px;
-  margin-top: 5px;
-  margin-bottom: 5px;
-}
-
-.confirm-button {
-  align-self: flex-end;
-  background-color: var(--sapphire);
-  color: white;
-  border-radius: 7px;
-  height: 33px;
-  font-size: 14px;
-  text-transform: none;
-  margin-top: 20px;
-  width: 89px;
-}
-
 .upload-message {
   font-size: 12px;
   margin: 10px;
